Moved LocalAudioDecoder
